PPS-62: Refactor Icon atom
