+++ src/components/atoms/Icon/index.js
@@ -1,7 +1,11 @@
 import React from 'react';
-
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { array, func, oneOfType, string } from 'prop-types';
+import {
+    array,
+    func,
+    oneOfType,
+    string,
+} from 'prop-types';
 import styled from 'styled-components';
 
 const StyledIcon = styled(FontAwesomeIcon)`
