+++ src/components/atoms/index.js
@@ -1,7 +1,9 @@
 import Button from './Button';
+import Icon from './Icon';
 import TextBox from './TextBox';
 
 export {
     Button,
+    Icon,
     TextBox,
 };
